archival: switch the archive `kind` argument from bytes to str
This is intertwined with the `mode`, which is also used with `open()`, so the
goal is to get that to str so that the `pycompat.open()` call can be dropped.
Start simple because I had issues where the generated archive in some tests had
1 different byte for some reason when making the full change, even though the
content was binary equivalent when doing a Folder Compare of the archives with
BeyondCompare4.

The key for the `archivers` map is left alone for now.  The classes that are
stored are only ever called with 2 args, so this should be safe.