run-tests: call heredoctest.py directly
Since we know exactly where it is, there is not need to use `python -m
heredoctest` to call it, we can just call that script.

We more the script in "testlib" to reduce the test directory clutter. We might
want to create a "tooling" directory in the future for this kind of utility.