run-tests: call heredoctest.py directly
Since we know exactly where it is, there is not need to use `python -m
heredoctest` to call it, we can just call that script.

We more the script in "testlib" to reduce the test directory clutter. We might
want to create a "tooling" directory in the future for this kind of utility.

--HG--
rename : tests/heredoctest.py => tests/testlib/heredoctest.py
extra : rebase_source : ed898bd090e430ee47228c2807f800ca060cf38b
extra : topic : further-python-dir