git: switch changelog.shortest from LIKE to GLOB to speed up matching
For whatever reason, LIKE is way slower than GLOB.  The slowdown is a
function of the number of commits in the repository.

The following data has been collected on a repository with approximately
1.2M commits.  The numbers are in milliseconds and they represent the
latency of the changelog.shortest function as measured by time.time().

The shortest function was invoked via `hg log -l50 -T '{node|shortest}\n'`.

   Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
  279.0   284.3   419.8   381.7   421.3   426.2 (LIKE)
0.08535 0.10437 0.12231 0.11919 0.12779 0.21291 (GLOB)