git: skip recursing into manifest subtrees that are the same
If during manifest comparison we are presented with identical tree objects,
then we are guaranteed that:

 1. the contents of these tree objects are the same
 2. the contents of any of their children are the same

This speeds up 'hg diff -c <hash>' considerably since unchanged subtrees are
are skipped.  For example, a commit that modified the README in the linux
kernel repo (~84k tracked files) now takes 1.1s instead of 15.7s.