pyupgrade: convert to new style classes
These were likely missed when the bulk of the classes got this change back in
642e31cb55f0.  These were rewritten using a hacked version of `pyupgrade` 3.19.1
that enabled only the `new_style_classes` fixer.  See 24ee91ba9aa8 for details.

It's probably better to fix the thirdparty and contrib code upstream, but these
are small changes, like those made locally in the past here.  I don't feel like
pulling down and testing potentially a lot of changes just to satisfy this
linter.