py3: drop redundant `u''` prefixes on string literals
Strings are unicode on Python 3.  These were rewritten by `pyupgrade`.

It's arguably better to fix the `contrib` stuff upstream and then re-vendor it,
but I don't feel like waiting for that, and then all of the regression testing
involved to get a minor improvement in the codebase.  It was last vendored 5
years ago, and will likely be a large change anyway to drop py2 support.  Also,
we've already made minor formatting changes to it locally.