typing: add some type annotations to the `wireprotoframing` module
`flushcommands()` and a few other generators got a slightly modified return type
after the pyupgrade series ending with 70a75d379daf.  This fixes those, and grabs
some other low and mid hanging fruit here.

Note that there are possibly two bugs detected here- in the `inputstream` and
`outputstream` classes, the `flush()` and `finish()` methods respectively have
surprising issues.  pytype flagged the latter when the return value was typed,
but it surprisingly doesn't see the lack of `flush()` on the `Decoder` class.
I wonder if either of these are the cause of the occasional odd-length string
error I see in the `hg serve` output from time to time.  In any event, punt
those changes down the road for now, and mark with a TODO.