contrib: move the MinGW download location to mercurial-scm.org
The old site has an invalid certificate, which causes the download to fail, and
derails the whole script.  It can still be manually downloaded from the old
site, but I had a local copy that passed the checksum verification, so let's
just host it like the Visual C for Python 2.7 compiler package in the same
location.