pyupgrade: drop `coding=UTF-8` comments
PEP-3120[1] (Python 3.0 in 2007) says that UTF-8 is the default encoding.  That
should be long enough ago that no reasonable editor would trip over this, and
certainly any supported version of Python won't.

The comments were probably harmless, but as `pyupgrade` has no mechanism to
disable this change, omitting this change makes it unusable as a code checking
tool, and makes it a pain to use occasionally to upgrade the source (since these
changes would need to be manually reverted).

[1] https://peps.python.org/pep-3120/