pyupgrade: drop redundant `open()` modes
On one hand, this gets rid of some py2 cruft (like `U` and `t` modes).  OTOH, I
like being explict.  But there's no way to disable individual transformations
(and this change is part of `_fix_tokens()`, which is even harder to selectively
use locally), and I'd rather have less noise when using `pyupgrade`.

Warning: `contrib/synthrepo.py` appears not to have been updated to py3, as it's
passing str to `error.Abort`.  But since it was opening the file in unicode on
py2, this change is a no-op.