tests: obfuscate Python version checks to subvert `pyupgrade`
When specifying `--py38-plus`, it reasonably deletes conditionals that check for
that version of python or earlier.  But we actually want the check to remain, so
use a simple obfuscation to preserve it.

--HG--
extra : rebase_source : 7e20d8064f860925ac4c0ac2c4c17ba46c83b747
extra : topic : pyupgrade-2