packaging: modernize (compat PEP 517) with less distutils and setup.py calls
- setup.py: less distutils imports and setuptools required

distutils is deprecated and one should import commands from setuptools to support
modern workflows depending on PEP 517 and 518.

Moreover, for Python >=3.12, distutils comes from setuptools. It corresponds to old and
unmaintain code that do not support PEP 517.

The PEP 517 frontends (pip, build, pipx, PDM, UV, etc.) are responsible for creating a
venv just for the build. The build dependencies (currently only setuptools) are specified
in the pyproject.toml file. Therefore, there is no reason to support building without
setuptools.

Calling directly setup.py is deprecated and we have to use a PEP 517 frontend.
For this commit we use pip with venv.

- run-tests.py: install with pip instead of direct call of setup.py

Mercurial is then built in an isolated environment.

- Makefile: use venv+pip instead of setup.py