pyoxidizer: fix building with PEP 517 support
I'm not sure when this broke, because f6b045910d82 explicitly says it's fixing
something, but a recent attempt to build pyoxidizer on Windows specifically
flagged `--no-use-pep517` as a problem.  Now that we have PEP 517 support for
building, this goes away and pyoxidizer 0.17.0 with rust 1.54.0 works again
(yeah, these should be updated).

Note that `py -3.9 run-tests.py --pyoxidized` fails to get the version info on a
zstd repo, and thus fails build and install hg to run the tests.  A zlib repo
works, and since there's a PR that changes obtaining the version info from
`setuptools-scm`, I'm not bothering to fix this.  With the zlib test run, the
following are broken:

    Failed test-commit-interactive.t: output changed
    Failed test-i18n.t: output changed
    Failed test-log.t:

The diffs all look to be related to not applying the requested i18n locale.  IDR
if this was the case when this was being tested 4 years ago (but I suspect it
was not).