hghave: byteify the `regexp` arg to `matchoutput()`
This started failing in CI because a str was passed.  All other callers already
pass bytes, and it's not clear why this started failing with a MR to use
`setuptools-scm` to derive the version when building.  Clearly the
`if checks["hg09"][0]()` prior to these usages were working before, and that's
how it was overlooked.  But that check apparently continued to work for me
locally in both Linux and Windows, as I can only hit this issue in CI.

--HG--
extra : rebase_source : 4a9a27ed9b53f69b4a4a0f5a04b22f7ea01137eb
extra : topic : setuptools-scm