streamclone: unbyteify string args to builtin Error classes
This avoids printing the error with a `b''` prefix in the case of `ValueError`.
The custom `ProgrammingError` class is special in that it won't do that, and can
take either bytes or str.  But there's no point in passing bytes when it is just
going to decode to str at runtime anyway.

--HG--
extra : rebase_source : 076e28f3e6e4b518019ed7830acd18b71b9e4b2c
extra : topic : pyupgrade-typing