typing: add some type annotations to the `merge.mergeresult` class
The generated type annotations around `filemap()` and `files()` were slightly
modified by the pyupgrade series culminating in 70a75d379daf.  This module is
way more complicated than the other changes, but these weren't too bad to figure
out.

The typing caught a trivial issue in `sparse`- it was passing an empty data list
to `addfile()` for the `ACTION_REMOVE` case, instead of a tuple or None.
`merge.manifestmerge()` calls this function with None for the data, so 1) it has
to be typed as optional, and 2) is safe to pass None in the sparse code.

--HG--
extra : rebase_source : 3d04878d993db76d947f933e662fb42d643f2970
extra : topic : pyupgrade-typing