typing: add type annotations for wireprotoframing encoder/decoder classes
This was low hanging fruit, and will point out an apparent programming bug.

The arguments are left untyped for now, because the internal compressors and
decompressors generally take a `Buffer` object, and return bytes.  I've had
issues around using `typing_extensions.Buffer` with pytype in the past, but also
the identity classes return the argument unmodified, so that will break the type
annotation.  We'd probably need an instance check, and a cast if it's not bytes.

Having to roll up all of the encoder and decoder types into a union type is a
bit annoying, but the protocol class can't be used in the map because it then
assumes that type will be instantiated.  I also couldn't get it to work with a
`TypeVar` bound to indicate that it's actually a subclass either.  Oh well.