usage: refactor usage.resources config parsing
This makes get_resource_profile less confusing. Before, RESOURCE_DEFAULT was a
special indirect value which meant to use DEFAULT_RESOURCE. I removed it while
preserving the same behavior.

The motivation for this is that the Rust implementation gets the default logic
wrong due to confusion about defaults. I will fix it in a follow-up commit.