hg-core: fix usage.resources default logic
This makes hg-core use ResourceProfileValue::Medium as the default. Before, it
used ResourceProfileValue::Default (now removed), which was not supposed to be a
real value, but rather an indirection meaning to use the default (medium).

The motivation for this is that my implementation of rhg annotate was slower on
some files than Python. This was because Python used the "Medium" profile by
default (enabling the revlog chunk cache), while Rust used the "Default" profile
(disabling the revlog chunk cache).