rust-manifest: add Manifestlog::inexact_data_delta_parents
This is similar to manifestctx.read_delta_parents(exact=False) in manifest.py.
It is useful to determine if a file was added in a changeset without
delta-resolving the entire manifest. I will use it for rhg annotate.

--HG--
extra : rebase_source : 64ca300e695e121623413d74525675cc5c738317
extra : topic : annotate