pyproject: add a note about possibly building more packages for macOS
Let's wait and see if anyone complains about this before proceeding.  But I
stumbled across this, and thought it might be useful to keep an eye on.  The
most relevant thing today is that cffi doesn't use Universal binaries, but it's
also been badly broken on macOS in the past, so maybe nobody is using it.

--HG--
extra : rebase_source : 20388be2e1f2862ac36ab90a328d92faad1d3a6a
extra : topic : cibuildwheels-pyproject