stream-clone-v2: turn the file chunking function into a class
We did this in two step: extracting the code to a function, then turning the function into a class for clarity. The introduction of the class adds a significant amount of boiler plate and complexity and make the previous move hard to follow.

--HG--
extra : rebase_source : 733a29ddb69f3ed3dace9ea2da7dbd1e7953a939
extra : topic : clonebundle-refactor-2