stream-clone-v2: use the pass through for threaded stream clone
The reader thread just want to iterate over chunk, so let it do just that. This avoid some silly layer of filechunkiter and chunk buffer.

Before this change, we had the following layers:

    bundle2part (iterator of chunks)
    → chunkbuffer (iterator of chunks to file-like)
    → filechunkiter (file-like to chunkgs)
    → _DataQueue (thread safe iterator of chunks)
    → chunkbuffer (iterator of chunks to file-like).

We now have removed the first layer of chunkbuffer + filechunkiter:

    bundle2part (iterator of chunks)
    → _DataQueue (thread safe iterator of chunks)
    → chunkbuffer (iterator of chunks to file-like).

This fix the output order change that was previously introduced.