rust-inner-revlog: move to `RwLock` instead of `RefCell`
We need to make `InnerRevlog` `Sync`, since we want to iterate on and read it
in parallel. An uncontended `RwLock` has no measurable overhead over a
`RefCell` in our contexts (to be verified in benchmarks, but I'm pretty sure).

This change assumes that writes/reads to/from the uncompressed chunk cache
can fail if contended, since it's just a cache, but the rest of operations
(which are on FileHandles) should wait for the lock and bubble up the panic
in case of lock poisoning. If a program tries to write from multiple threads
to a revlog, it had better have good reasons, so I'm not too worried.

--HG--
extra : rebase_source : 431f9d405000a0a9447e219ab46ef64054b7cec9
extra : topic : annotate