chunkbuffer: provide a "__iter__" pass through
If the consumer just want to iterate over chunk of arbitrary size, the chunk
buffer is just an overhead. So we provide a way to bypass it in such cases.

It is put to use in the next changesets.

--HG--
extra : rebase_source : 807083a87cc7ef577b432acf904806b037f7cb59
extra : topic : clonebundle-refactor-2