pytype-docker: make sure the ~/.local/ directory exists
In the case the base image is already properly setup with pytype, the pip
install will not do anything and not create the ~/.local directory. This lead
the later chmod to fail. It seem like a good idea to keep that pip call so I
just create the directory first.