stream-clone-v2: centralize preparation for streamed files
The vfs will do some operation to prepare the creation of the streamed file. It
creates the necessary directories and it adds the file to the fncache. The
directory creation is much more expensive than it should, and the same
directory are being checked over and over. And the other part, like "adding the
file to the fncache" increase the amount of GIL holding time.

So we add a function to the vfs to do such preparation, and we centralise it in
the "parsing" thread.  This is done so that each directory are only "created" once.

This does not yield speed up yet, because the writer thread still go through the
same slow logic. This will be dealt with in the next changesets.