run-test: do not mark test as timed out when they returned fine
This while look break when the process finish or when the time out is reached.
So we need to check why we existed the loop.

Not sure how it has not created more problem so far.

--HG--
extra : rebase_source : 13c02865d81729958df2ef9bb07a6eeca5d1120b
extra : topic : timeout