rust-pyo3-dirstate: making bytes slices in core Sync
For the purposes of providing PyO3 bindings, the data fields that will be
exposed to Python  have to be `Sync`, hence that is the case of
`OwningDirstateMap` and its `owner` field.

We had to do something similar for the PyO3 bindings of `revlog`.

In this case, it forces us to adapt the `Deref` wrapper of `PyBytes` used
in `hg-cpython`, because it must itself now be `Sync` and raw pointers are
not.

This looks even uglier than it used to, but it does not matter much, because
our ultimate goal is to remove the rust-cpython bindings altogether.

--HG--
extra : rebase_source : 2334c176aeea24105dc9cd584e7dfb79d43e5fa4
extra : topic : pyo3.dirstate