rust-pyo3: making PyBytesDeref Sync
This parallels what we have done earlier with the version in rust-cpython.
It was not needed until now because the previous use of this struct
was in `InnerRevlog` for the single revision cache, and which is shielded in
`hg-core` behind a `Mutex`.

But `DirstateMap` is needing its `owner` to be `Sync` more directly.
In turn it means that maybe we could remove the `Mutex` of `hg-core`.