rust-pyo3-dirsate: DirstateMap definition and constructors
With rust-cpython, using `Option<T>` in the Rust signature would
translate automatically into a Python keyword argument, leading
to unpleasantness on the Rust side in some cases (not here, though).
with PyO3, though the signature can be specified explicitly (positional
in this case) and there is even a warning that the implicit keywords
style is deprecated.