rust-pyo3-dirstate: DirstateItem class
The rust-cpython version was using `Cell`, which is not `Sync`, hence
we replace it with the pervasive `RwLock`. This brings in turn some
minor simplification, as we can access a mutable reference to the inner
data directly, and do not need the `update()` method (a shortcut to
perform the `get`/`set` dance of the `Cell`).

Tighter signature control in PyO3 means that `set_fallback_exec` can
probably be simplified, which we might do once everything is validated
by the tests.