dispatch: allow global options with a '-' in the long name
The handful of current global commands are one word, so this hasn't been
necessary.  Unfortunately, the `fancyopts` class transforms `--long-arg-name`
into `long_arg_name` when it populates the `options` list used here.  That
results in a KeyError when iterating over `commands.globalopts` and doing a
lookup with the dash-based name that is required there, in order to have the
dash-based name in the user interface.