clone: explicitly steal lock instead of assigning previous lock
The issue with reusing the lock from another repository is that various internal
state are no longer correct, the main example of that is the dirstate, captured
in the wlock to make sure it is written (when needed) on lock release. So
instead, we create a proper lock from the repository, but "stealing" the on-disk
lock from the previous object.

This is a bit weird, but less than the previous situation.