test-config: handle multiple path separators on Windows
This isn't a current problem because the only thing currently passed is in the
form `"file1:file2"`.  When I had to patch in a 3rd file to load an extension,
that caused the config to not be read from those two files, because only the
first `:` was replaced.  Let's save somebody else time in the future if they add
another file.