overlayctx: build a memctx that can access non touched files
So far, the memctx was very optimized for being committed: so it did not need to
access any untouched files.

However the way we uses it in evolve to diff against more arbitrary changes
requires to be able to access to any file content as any other ctx.

This is now the case.

--HG--
extra : amend_source : 4981b667f80d53c4a4ee637643b649e9e6431832
extra : topic : diffchanged