refactor: split `bundlerepo.getremotechanges` into three pieces
Split the function in three and introduce an explicit class that tracks state.

The goal here is to improve error handling, in particular so it's easy to
wrap things in a try/with without having to grow an already-significant
indentation level.

The code can probably be cleaned up further, but I don't want to bite off
a piece too large to chew.