rust-pyo3-dirstate: CopyMap without proxy methods
It turns out that it is fairly easy to call the core `DirstateMap`
methods from a `CopyMap` PyO3 object, hence we do not need to introduce
proxy methods on the FFI `DirstateMap` as was done in `rust-cpython`.
It is more straightforward, and is somewhat easier to make with PyO3
because we have all these intermediates between the Rust and Python layers
(`Bound`, `Py<T>`).

From the Python side, a difference with the `hg-cpython` version is that
`CopyMap` would be instantiable also directly as `CopyMap(dirstate_map)`.
Reproducing the `from_inner`, with refcount increase the responsibility in the
`DirstateMap.copymap()` method would have been less natural (as there is
no need with the separate `from_inner` in PyO3 constructors) and somehow
we felt it was sounder to force the ref cloning from `CopyMap`.

--HG--
extra : rebase_source : 0bb69bbb24ef2e54177e243db2df5a954be49766
extra : topic : pyo3.dirstate