rust-pyo3-sharedref: macro to define Python iterators
This convenience will help reducing the boilerplate for one of
the most common use cases of the `pyo3_sharedref` crate.

It is an improved adaptation of a similar macro that was in `hg-cpython`
(but not in `rust-cpython`). The improvement is that it takes care of
the constructor, sparing the `share_map` to the caller.

--HG--
extra : rebase_source : 5ae0771b98c1df1f206ee058f1ca3a0b0c49da0c
extra : topic : pyo3.dirstate