rust-pyo3-sharedref: macro to define Python iterators
This convenience will help reducing the boilerplate for one of
the most common use cases of the `pyo3_sharedref` crate.

It is an improved adaptation of a similar macro that was in `hg-cpython`
(but not in `rust-cpython`). The improvement is that it takes care of
the constructor, sparing the `share_map` to the caller.