sparse: add test for status --rev --rev and --change
Sparse only applies to the working copy, not history, so it should be ignored
for status --rev --rev. Python is correct here but rhg has a bug, introduced in
136e74c2bf8f917cf963dc1b594b6a90f8e5f627. I will fix it in the next changeset.

(On the other hand, status --rev --rev *should* respect the narrowspec. There is
already a test for this at the bottom of tests/test-narrow-patch.t.)