rust-pyo3-dirstate: Dirs methods not involving iterators
We hesitated on this one: was it worth it to introduce the `with_inner_*`
helpers? Finally quickness and uniformity decided we should.

At this point it would be interesting to do something in `pyo3-sharedref`
to provide them automatically, but some callers will need the self reference
(as in `revlog`), some would not.

--HG--
extra : rebase_source : 8a6ca073b23717fddec2671aae1dae87b69d7510
extra : topic : pyo3.dirstate