sshpeer: fix deadlock on short writes
This commit makes the `sshpeer.doublepipe` object retry on short write,
which fixes a deadlock in hg client-server communication, in
particular when client needs to send a large message.

Apparently Mercurial relies on `write` method on file objects
to never return short writes, without checking, which leads to deadlocks.

This work fine when the underlying file object is buffered,
since buffered writers never return short writes.
(why buffering has anything to do with this I don't know, but ok)

It breaks horribly with raw IO, which happens to be used in sshpeer.