dispatch: restore `--config` can't be abbreviated message in case of ambiguity
This feels kind of gross, in part because it's a surprising amount of code for
a specific edge case, and in part because it assumes some implementation
details.  But I don't see a better way, since `getopt.long_has_args()` will
create a list of all known long options that start with the current option, and
raise errors of this type if the list is either empty or has more than one
element.  There's not a programmatic way to tell the difference between these
two aside from parsing the message, so we replicate the check here and raise
the expected exception in the case of ambiguity.  Otherwise, the original
expection is re-raised.