annotate: add test for ancestor check starting revision
Python hg annotate currently calls .introfilectx() and uses that for all
ancestor checks. This is an unnecessary extra step, and it results in unstable
output as shown in the test. It should instead behave like rhg annotate and use
the starting revision passed on the command line for ancestor checks.

This was discovered here:
https://foss.heptapod.net/mercurial/mercurial-devel/-/merge_requests/1179#note_380887