rhg-annotate: support whitespace options
This adds support to rhg annotate for all the whitespace options:

    -w, --ignore-all-space
    -b, --ignore-space-change
    -B, --ignore-blank-lines
    -Z, --ignore-space-at-eol

Note that --ignore-blank-lines has no effect on annotate so it is ignored. You
can see this in dagop.py _annotepair which only checks if blocks are '=' or not,
whereas the effect of --ignore-blank-lines is to change some '!' into '~'.

When the other 3 are combined, we use the strongest option since -w implies -b
and -b implies -Z. This is not explicit in the Python implementation, but I have
verified that's how it behaves.

--HG--
extra : rebase_source : 1906887671560a2eb6c285ac54867a6aea03cc2e
extra : topic : annotate-whitespace