extensions: allow wrapping a function with a bytes name again
This backs out 7b837fabc990 and eda075d7b2ac.

The cleanup to transition away from bytes broke JavaHg, which is used by hosting
software SCM Manager.  I'm not sure who maintains JavaHg these days, so I'm not
sure how to get a build out and get everyone to upgrade.  Let's revert the
change for a quicker fix (and fix an unknown number of other unmaintained
extensions), and rely on pytype to flag any potential issues.  It's also
trivial, so let's not wait until the next feature release (with the added burden
for packagers this cycle in dealing with the `setup.py` changes).