typing: add an interface for url
This is fairly simple. This is a dependency for the "path" class in the same
module, that is itself imported by `mercurial.interfaces.repository`. So we get
it out of the way first.

--HG--
extra : rebase_source : 1206fc659a24f2c9d79398bc255fb5e6e5b25cde
extra : topic : pytype-cycle