cleanup: remove the sqlitestore extension
The idea of having a clear storage interface and a small extension to test it
was not bad in theory. However this never materialize to anything concrete, and
nobody has touched this extensions for over 6 years. So it is mainly getting in
the way of other works.

Incidently, the clarification of interface is seeing progress thanks to the Matt
Harbison typing effort and the slow rewrite into rust of core element led by
Octobus.

So I suggest we drop this extension, freeing energy for other improvements.