revlog: introduce a REVIDX_HASMETA flag
We don't actually store it yet, but we can use it to push the logic for
swapping parent further down.

The final logic will not be as simple because we need to handle the network
exchange efficiently, but this demonstrate that we can mostly dead with this
inside the filelog itself without too much troubles.

--HG--
extra : rebase_source : f0873776bbfc6a6cab65870024f10857d58a95c3
extra : topic : fl-meta-flag