typing: add a cast to `mercurial.patch` to unconfuse pytype 2024.10.11
I've seen this in other, more modern versions of pytype than what we use in CI-
it thinks the return of `get_payload()` is `Message | bytes`, and then complains
about `if not payload.endswith(b'\n')` below, saying `Message` doesn't have that
function.  That's correct, but according to the table in `get_payload()` that
documents what the method returns, the only possible values are `bytes | None`
when `decode=True` and the `i` argument isn't passed.  And `None` isn't possible
if we trust the `Content-Type` header.

While here, I added a type hint for the `part` variable, since PyCharm couldn't
figure out what was going on here, and that made navigating to the code
impossible.