index: remember the generaldelta config instead of getting it from the revlog
The code (especially the Rust code) was jumping in 4 dimensions to make sense
of what was going on because it wrongly assumed that we needed to somehow
be able to ask a generaldelta index for a non-generaldelta delta chain, which
doesn't make any sense.

Removing the cargo-culted/vestigial code, this is cleaner and less confusing.

--HG--
extra : rebase_source : 96b6ade26487b427c6d1017113f9ee84e627a54e
extra : topic : index.deltachain-signature